client: make isolateserver_smoke_test runnable on python3
Bug: 1010816
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/2277551
Commit-Queue: Junji Watanabe <jwata@google.com>
Auto-Submit: Takuto Ikuta <tikuta@chromium.org>

+++ client/tests/isolateserver_smoke_test.py
@@ -41,8 +41,8 @@ class IsolateServerArchiveSmokeTest(unittest.TestCase):
     # before being uploaded.
     # TODO(maruel): This should not be leaked to the client. It's a
     # transport/storage detail.
-    self.namespace = ('temporary' + str(long(time.time())).split('.', 1)[0]
-                      + '-gzip')
+    self.namespace = ('temporary' + str(int(time.time())).split('.', 1)[0] +
+                      '-gzip')
     self.tempdir = tempfile.mkdtemp(prefix=u'isolateserver')
     self.rootdir = os.path.join(self.tempdir, 'rootdir')
     self.test_data = os.path.join(self.tempdir, 'test_data')
